% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{axis}
\alias{axis}
\title{Set a font for a chart axis}
\usage{
axis()
}
\value{
font properties
}
\description{
Set parameters for font that will be used in a chart axis, e.g. when using package plotly.
}
\examples{
\dontrun{
  xaxisProperties <- list(title = "Date", titlefont = chartfonts::axis()), <rest_of_parameters>)
}
}
\author{
from-to: 2020-10-13-
}
\keyword{font}
