% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hover.R
\name{hover}
\alias{hover}
\title{Set a font for a mouse hover}
\usage{
hover()
}
\value{
font properties
}
\description{
Set parameters for font that will be used in a chart after mouse hover, e.g. when using package plotly.
}
\examples{
\dontrun{
  hoverlabel = list(font = chartfonts::hover(), align = "left")
}
}
\author{
from-to: 2020-10-13-
}
\keyword{font}
