% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\name{ticker}
\alias{ticker}
\title{Set a font for a chart ticker}
\usage{
ticker()
}
\value{
font properties
}
\description{
Set parameters for font that will be used in a chart ticker, e.g. when using package plotly.
}
\examples{
\dontrun{
  xaxisProperties <- list(
    title = "Date",  # axis title
    titlefont = fontsUDP::DisplayFonts("axis"),
    showticklabels = T,
    tickformat = "\%d \%b \%y",
    tickangle = 45,
    tickfont = chartfonts::ticker(),
    <rest_of_parameters>
  )
}
}
\author{
from-to: 2020-10-13-
}
\keyword{font}
