% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{title}
\alias{title}
\title{Set a font for a chart title}
\usage{
title()
}
\value{
font properties
}
\description{
Set parameters for font that will be used in a chart title, e.g. when using package plotly.
}
\examples{
\dontrun{
  plotly::layout(title = "Title text", font = chartfonts::title(), <rest_of_parameters>)
}
}
\author{
from-to: 2020-10-13-
}
\keyword{font}
